Test Model Loader

+++ Source/Core/Loader/ERS_ModelLoader/ERS_ModelLoader.cpp
@@ -72,7 +72,7 @@ void ERS_CLASS_ModelLoader::LoadModel(long AssetID, std::shared_ptr<ERS_OBJECT_M
 
     std::shared_ptr<ERS_STRUCT_IOData> ModelData = std::make_shared<ERS_STRUCT_IOData>();
     SystemUtils_->ERS_IOSubsystem_->ReadAsset(AssetID, ModelData);
-    const aiScene* Scene = Importer.ReadFileFromMemory(static_cast<const void*>(ModelData->Data.get()), ModelData->Size_B, aiProcess_Triangulate | aiProcess_GenSmoothNormals | aiProcess_FlipUVs | aiProcess_CalcTangentSpace, "");
+    const aiScene* Scene = Importer.ReadFileFromMemory(static_cast<const void*>(ModelData->Data.get()), (int)ModelData->Size_B, aiProcess_Triangulate | aiProcess_GenSmoothNormals | aiProcess_FlipUVs | aiProcess_CalcTangentSpace, "");
 
     // Log Errors
     if (!Scene || Scene->mFlags & AI_SCENE_FLAGS_INCOMPLETE || !Scene->mRootNode) {
